factor out histogram.js from explorer.js.
